Downloads the last 24 hours of points and plots them
The data source alerts the viewcontrollers when it has data. In the
future this can be used to grab new data.

We also introduce a class to manage the Shinobi license key too,
which stores it in a plist, which isn't in the repo. Instructions are in
the VPYShinobiLicense.m file.