Updates license key code options
We now always provide a ConAirSettings.plist, based on a sample one
which doesn't leave my machine. The githooks are used to swap out the
one with the correct value with the sample one.

Also refactored the ShinobiLicense class to fail with a more intelligent
error when a licence isn't available.